added try catch and added param to error logging.

+++ MicroApplication/BaseLibrary/Controls/Forms/ControlReader.cs
@@ -84,7 +84,7 @@
         {
             string value = GetControl(appControl, ControlValues)?.GetFirstValue();
             if (!C.IsNull(value))
-                return Convert.ToInt16(value);
+                return Convert.ToInt32(value);
             return 0;
         }
         public static double GetControlFirstValueAsDouble(AppControl appControl, List<ControlValue> ControlValues)
